feat(inputQueue): 实现输入队列的后台运行和状态持久化
- 新增后台输入队列运行器，支持队列项自动出队并执行请求发送
- 通过本地存储持久化输入队列状态，实现跨标签页数据同步
- 使用锁机制确保输入队列状态读写和发送操作的互斥，防止竞态条件
- 当队列空闲时会自动释放所有权，支持所有权心跳保持和刷新
- 输入队列错误恢复机制，失败任务支持重新入队与恢复
- 优化聊天控制器逻辑，拆分输入队列控制相关功能，提升代码可维护性
- 移除旧版输入队列状态管理，改用独立hook管理队列生命周期
- 新增相关单元测试覆盖输入队列状态持久化和恢复逻辑
